Refactor relative imports to use absolute paths, update routes
