+++ frontend/src/modules/Shares/components/ShareOutboxList.js
@@ -5,9 +5,9 @@ import { Helmet } from 'react-helmet-async';
 import { Defaults, Pager, useSettings } from 'design';
 import { SET_ERROR, useDispatch } from 'globalErrors';
 import { getShareRequestsFromMe, useClient } from 'services';
-import ShareOutboxListItem from './ShareOutboxListItem';
+import { ShareOutboxListItem } from '../components';
 
-const ShareOutboxList = () => {
+export const ShareOutboxList = () => {
   const dispatch = useDispatch();
   const [items, setItems] = useState(Defaults.pagedResponse);
   const [filter, setFilter] = useState(Defaults.filter);
@@ -88,4 +88,3 @@ const ShareOutboxList = () => {
   );
 };
 
-export default ShareOutboxList;
